requirements.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1
pandas==2.1.4
python-dotenv==1.0.0
